+++ samples/Demo/Beef.Demo.Database/Migrations/20210127-175235-create-democdc-person2outbox.sql
@@ -5,8 +5,8 @@ CREATE TABLE [DemoCdc].[Person2Outbox] (
 
   [OutboxId] INT IDENTITY (1, 1) NOT NULL PRIMARY KEY CLUSTERED ([OutboxId] ASC),
   [CreatedDate] DATETIME NOT NULL,
-  [Person2MinLsn] BINARY(10) NOT NULL,  -- Primary table: Demo.Person2
-  [Person2MaxLsn] BINARY(10) NOT NULL,
+  [Person2MinLsn] BINARY(10) NULL,  -- Primary table: Demo.Person2
+  [Person2MaxLsn] BINARY(10) NULL,
   [IsComplete] BIT NOT NULL,
   [CompletedDate] DATETIME NULL,
   [HasDataLoss] BIT NOT NULL